# internal/api/client/api/swagger.yaml
openapi: 3.0.3
info:
  title: Frank Auth Server API
  description: "Frank is a comprehensive authentication and authorization server that\
    \ provides OAuth2, passwordless login, MFA, \npasskeys, SSO, enterprise SSO, webhooks,\
    \ organizations, and API keys for machine-to-machine authentication.\n"
  contact:
    name: Frank Auth Server
    url: https://github.com/juicycleff/frank
  version: 1.0.0
servers:
- url: https://auth.example.com/api/v1
  description: Production server
- url: http://localhost:8080/api/v1
  description: Local development server
...
paths:
...
  /auth/sso/providers:
    get:
      tags:
      - SSO
      summary: List SSO providers
      description: Get the list of available SSO providers
      operationId: ssoProvidersList
      parameters:
      - name: organization_id
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: Available SSO providers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/inline_response_200_20"
  /auth/sso/providers/{provider}:
    get:
      tags:
      - SSO
      summary: Authenticate with SSO provider
      description: Start SSO authentication with a provider
      operationId: ssoProviderAuth
      parameters:
      - name: provider
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: redirect_uri
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
          format: uri
      responses:
        "302":
          description: Redirect to provider
          headers:
            Location:
              style: simple
              explode: false
              schema:
                type: string
                format: uri
        "400":
          description: Invalid provider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/sso/callback/{provider}:
    get:
      tags:
      - SSO
      summary: SSO callback
      description: Handle SSO callback from a provider
      operationId: ssoProviderCallback
      parameters:
      - name: provider
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: code
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: SAMLResponse
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: state
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/inline_response_200_19"
        "302":
          description: Redirect to application
          headers:
            Location:
              style: simple
              explode: false
              schema:
                type: string
                format: uri
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/Error_error"
...
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        first_name:
          type: string
        last_name:
          type: string
        phone_number:
          type: string
        profile_image_url:
          type: string
          format: uri
        locale:
          type: string
        email_verified:
          type: boolean
        active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties: true
        last_login_at:
          type: string
          format: date-time
      example:
        metadata: ""
        last_login_at: 2000-01-23T04:56:07.000+00:00
        email_verified: true
        last_name: last_name
        active: true
        created_at: 2000-01-23T04:56:07.000+00:00
        profile_image_url: http://example.com/aeiou
        locale: locale
        updated_at: 2000-01-23T04:56:07.000+00:00
        phone_number: phone_number
        id: id
        first_name: first_name
        email: ""
...
    SSOProvider:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum:
          - saml
          - oidc
        logo_url:
          type: string
          format: uri
        domains:
          type: array
          items:
            type: string
        organization_id:
          type: string
      example:
        logo_url: http://example.com/aeiou
        organization_id: organization_id
        name: name
        domains:
        - domains
        - domains
        id: id
        type: saml
...
    inline_response_200_19:
      type: object
      properties:
        message:
          type: string
        user:
          $ref: "#/components/schemas/User"
      example:
        message: message
        user:
          metadata: ""
          last_login_at: 2000-01-23T04:56:07.000+00:00
          email_verified: true
          last_name: last_name
          active: true
          created_at: 2000-01-23T04:56:07.000+00:00
          profile_image_url: http://example.com/aeiou
          locale: locale
          updated_at: 2000-01-23T04:56:07.000+00:00
          phone_number: phone_number
          id: id
          first_name: first_name
          email: ""
    inline_response_200_20:
      type: object
      properties:
        providers:
          type: array
          items:
            $ref: "#/components/schemas/SSOProvider"
      example:
        providers:
        - logo_url: http://example.com/aeiou
          organization_id: organization_id
          name: name
          domains:
          - domains
          - domains
          id: id
          type: saml
        - logo_url: http://example.com/aeiou
          organization_id: organization_id
          name: name
          domains:
          - domains
          - domains
          id: id
          type: saml
...
    Error_error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        data:
          type: object
          additionalProperties: true
...
  securitySchemes:
    bearerAuth:
      type: http
      description: JWT token obtained from the /auth/login endpoint
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      description: API key for machine-to-machine authentication
      name: X-API-Key
      in: header